Updated tests to comply with dynamic return type
A couple of tests that explicitly invoked a Route and asserted
on the return. There was also a case where an implicit cast, from
null to Response, could happen inside the route action, before it
was invoked.

Returning null from a route or returning "foo.bar" where bar is null
causes different behaviors in C# when the return type has an implicit
cast operator that can be applied. With the change to dynamic return,
it will always return a proper null, thus some tests needed to be
setup differently