Introduced the IRequestDispatcher
It supports the convention of turning a path file extension into an
accept header, using the ExtensionMappings of the IResponseProcessor
implementations that are available. The accept headers will get a
quality of 1.1 so they take presedence.

Failing to match the extension less path, with the added header(s),
will result in a second attempt to match a route using the original
path and headers