render child requests in razor views
This is a proof of concept on rendering child requests within views.
This allows for more modular view pieces than you get with Partial
views, as the parent doesn't have to worry about constructing the model
for the child view piece.  To see the proof in action, launch the
Nancy.Demo.Hosting.Aspnet project and browse to the /viewwithchildview
route.